Create  Total_Rainwater_Brute_Force.cpp
What this program does is it takes an array as input from the user. The array contains the heights of buildings and the array size represents the total number of buildings. It tell how many units of rainwater will be stored between the buildings if rained assuming that they are 2D.

For this purpose it finds number of units of rainwater present on each building by finding its left most maximum element and right most maximum element, then it does some calculations and return the total number of units of rainwater stored between the buildings.